+++ src/index.tsx
@@ -1,9 +1,8 @@
-/* @refresh reload */
 import { render } from "solid-js/web";
 
 import App from "./App";
 import "virtual:uno.css";
-import "./tailwind-compat.css";
-import "./icons.css"
+import "./styles/tailwind-compat.css";
+import "./styles/icons.css"
 
 render(() => <App />, document.getElementById("root") as HTMLElement);
