+++ src/types/FollowRecord.tsx
@@ -1,0 +1,11 @@
+type FollowRecord = {
+  did: string;
+  handle: string;
+  uri: string;
+  status: RepoStatus;
+  status_label: string;
+  toDelete: boolean;
+  visible: boolean;
+};
+
+export default FollowRecord;
